move rgb to utils
