+++ src/routes/+page.svelte
@@ -12,7 +12,7 @@
     import { map, draw } from '../store/map'
     import { EMPTY_POLYGON_RGB } from '../lib/gl_draw_styles.js'
     import { DeleteClickedZone } from '../lib/custom_delete.js'
-    import { getClickPoint, findLeftTopY, findLefTopX, getObjectSizeWithStroke, UUIDv4 } from '../lib/utils'
+    import { getClickPoint, findLeftTopY, findLefTopX, getObjectSizeWithStroke, UUIDv4, getRandomRGB, rgba2array } from '../lib/utils'
 	import type { Polygon } from 'geojson';
 
     const { apiURL } = apiUrlStore
@@ -664,27 +664,6 @@
             .catch((error) => {
                 console.log('Error on replacing data', error)
             })
-    }
-
-    function getRandomRGB() {
-        // https://stackoverflow.com/a/23095731/6026885
-        const num = Math.round(0xffffff * Math.random());
-        const r = num >> 16;
-        const g = num >> 8 & 255;
-        const b = num & 255;
-        return 'rgb(' + r + ', ' + g + ', ' + b + ')';
-    }
-
-    const rgba2array = (rgbValue?: string): [number, number, number] => {
-        if (!rgbValue) {
-            return [0, 0, 0];
-        }
-        // https://stackoverflow.com/a/34980657/6026885
-        const match = rgbValue.match(/rgba?\((\d{1,3}), ?(\d{1,3}), ?(\d{1,3})\)?(?:, ?(\d(?:\.\d?))\))?/);
-        if (!match) {
-            return [0, 0, 0];
-        }
-        return [parseInt(match[1]), parseInt(match[2]), parseInt(match[3])]
     }
 
     // define a function that can locate the controls.
